(feat) Enhancements to the active visits table
This commit makes the following improvements to the active visits table:

- Rename the `ActiveVisitsListTable` component to `ActiveVisitsTable`.
- Move data fetching logic from the queue metrics resource to a new `ActiveVisitsTable` resource. There isn't sufficient overlap or code reuse between the two resources to warrant colocation.
- Change the default table size to `compact`.
- Show tooltips for visit priority when a note is present.
