+++ packages/esm-outpatient-app/src/active-visits/active-visits-table.scss
@@ -7,12 +7,12 @@
   background-color: $ui-01;
 }
 
-.activeVisitsListContainer {
+.activeVisitsTableContainer {
   width: 100%;
   height: 100%;
 }
 
-.activeVisitsListHeaderContainer {
+.activeVisitsTableHeaderContainer {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -37,6 +37,10 @@
 .switch {
    width: 7.875rem;
    height: 2rem;
+}
+
+.tooltip :global(.bx--tooltip__trigger.bx--tooltip__trigger--definition) {
+  border-bottom: none;
 }
 
 .priorityTag {
@@ -143,13 +147,3 @@
     border-bottom: 0.375rem solid var(--brand-03);
   }
 }
-
-.priorityTooltip {
-  @include carbon--type-style("body-short-02");
-  padding: 0.25rem; 
-}
-
-:global(.bx--tooltip--definition .bx--tooltip__trigger){
-  border-bottom: none;
-}
-
